# api/openapi.yaml
openapi: 3.1.0
info:
  title: Concord API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    Read and prepare-only access to Concord's FXRP-backed syndicated capital
    facility. The API projects authoritative onchain state and verified FCC
    evidence; it does not custody assets or replace Accord authority.
servers:
  - url: https://api.concord.invalid
    description: Placeholder; use the local CLI or an explicitly configured deployment URL.
...
security: []
...
paths:
...
  /v1/transactions/prepare:
    post:
      tags: [Transaction intents]
      summary: Prepare unsigned transaction calldata
      operationId: prepareTransaction
      description: >-
        Builds unsigned calldata only. The caller must review, authorize, sign,
        broadcast, and wait for the chain receipt. The API never returns a
        claim that the action succeeded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrepareTransactionRequest'
      responses:
        '200':
          description: Unsigned transaction intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionIntentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
components:
...
  responses:
...
    InvalidRequest:
      description: Invalid identifier or action arguments
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Bytes32:
      type: string
      pattern: '^0x[0-9a-fA-F]{64}$'
      description: Hex-encoded Solidity bytes32 value.
    Address:
      type: string
      pattern: '^0x[0-9a-fA-F]{40}$'
    DecimalString:
      type: string
      pattern: '^[0-9]+$'
      description: Non-negative integer in base-10 notation; never a floating-point JSON number.
    Observation:
      type: object
      required: [status, source, network, chainId, observedAt]
      properties:
        status:
          type: string
          enum: [observed, partial, not_observed]
        source:
          type: string
          enum: [onchain, fcc_evidence, derived]
        network: { type: string, example: coston2 }
        chainId: { type: integer, example: 114 }
        blockNumber:
          $ref: '#/components/schemas/DecimalString'
        observedAt:
          type: string
          format: date-time
        warning: { type: string }
    Meta:
      type: object
      required: [observation]
      properties:
        observation: { $ref: '#/components/schemas/Observation' }
        apiVersion: { type: string, example: v1 }
...
    PrepareTransactionRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [create_root, lock_collateral, approve_asset, fund_child, draw, repay]
        rootAccordId: { $ref: '#/components/schemas/Bytes32' }
        childAccordId: { $ref: '#/components/schemas/Bytes32' }
        drawId: { $ref: '#/components/schemas/Bytes32' }
        targetCapacity: { $ref: '#/components/schemas/DecimalString' }
        amount: { $ref: '#/components/schemas/DecimalString' }
        validUntilUnix: { $ref: '#/components/schemas/DecimalString' }
        policyHash: { $ref: '#/components/schemas/Bytes32' }
        asset: { $ref: '#/components/schemas/Address' }
        spender: { $ref: '#/components/schemas/Address' }
        actor:
          type: string
          enum: [treasury, provider, institution, agent]
    TransactionIntent:
      type: object
      required: [action, chainId, to, data, value, summary, requiresExplicitApproval, preconditions]
      properties:
        action: { type: string }
        chainId: { type: integer, example: 114 }
        to: { $ref: '#/components/schemas/Address' }
        data: { type: string, pattern: '^0x[0-9a-fA-F]*$' }
        value: { $ref: '#/components/schemas/DecimalString' }
        summary: { type: string }
        requiresExplicitApproval: { type: boolean, const: true }
        preconditions:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
...
    TransactionIntentResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/TransactionIntent' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
        meta: { $ref: '#/components/schemas/Meta' }